// src/types/style.d.ts
export type FlexDirection = 'column' | 'row' | 'row-reverse' | 'column-reverse';
export type FlexJustify =
  | 'flex-start'
  | 'flex-end'
  | 'space-between'
  | 'space-around'
  | 'space-evenly'
  | 'center';
export type FlexAlign =
  | 'flex-start'
  | 'flex-end'
  | 'baseline'
  | 'stretch'
  | 'center';
export type CommonSizeType = 'small' | 'medium' | 'large';